Fixed bug with index. Now deleting correctly.
There was a bug in the previous push that didn't
delete the corret item in myLibrary. That's now
fixed by using the book title to find the index
rather than setting a property on item creation.
The bug was that the index changed whereas the
property didn't. Last step is a toggle for read
and not read.